Sending the project for review
Major changes:
Restructured the files
Created TransactionTableRow component to reduce the size and complexity of TransactionTable component
WalletScreen --> OneWalletScreen
MyWalletsScreen is now a higher order component to reduce code duplication
Moved utility functions used in multiple components to utils in order to reduce code duplication
Extracted onSubmit functions from forms for better readability

Style changes:
Added animations to Wallet and TransactionTableRow
Edit, delete buttons are visible only if hovered on the parent element
Dialogs have now consistent styling
Check the user's color scheme preference and select the theme accordingly

Other:
Removed ESLint disables
