+++ zarofeladat/src/components/ConfirmDialog.js
@@ -7,7 +7,7 @@ export default function ConfirmDialog({
 }) {
   return (
     <Dialog open={open} onClose={handleClose} maxWidth="lg">
-      <DialogTitle DialogTitle variant="h5" textAlign="center" fontWeight={500}>{title}</DialogTitle>
+      <DialogTitle variant="h5" textAlign="center" fontWeight={500}>{title}</DialogTitle>
       <DialogActions sx={{ pl: 3, pr: 3, pb: 3 }}>
         <Button variant="contained" fullWidth onClick={handleEvent}>Yes</Button>
         <Button variant="contained" color="error" fullWidth onClick={handleClose}>Cancel</Button>
