+++ zarofeladat/src/screens/onewallet/hook/useTransactions.js
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { AXIOS_METHOD, doApiCall } from './useApi';
+import { AXIOS_METHOD, doApiCall } from '../../../hooks/useApi';
 
 export default function useTransactions(walletId, limit = 5) {
   const [transactions, setTransactions] = useState(false);
