+++ zarofeladat/src/screens/wallets/MyWalletsScreen.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { useAuth } from '../../hooks/useAuth';
+import WalletsScreen from './WalletsScreen';
+
+export default function MyWalletsScreen() {
+  const { sessionUser } = useAuth();
+
+  return (
+    <WalletsScreen filterBy={(d) => d.created_by.name === sessionUser.name} />
+  );
+}
