+++ zarofeladat/src/screens/wallets/WalletsScreen.js
@@ -1,19 +1,23 @@
-/* eslint-disable react/jsx-no-bind */
 import {
   Button, Grid, LinearProgress, Typography,
 } from '@mui/material';
 import { Container } from '@mui/system';
 import { useState } from 'react';
-import MenuBar from '../components/MenuBar';
-import Wallet from '../components/Wallet';
-import AddWalletDialog from '../Dialogs/AddWalletDialog';
-import { AXIOS_METHOD, doApiCall, useApi } from '../hooks/useApi';
-import { useAuth } from '../hooks/useAuth';
+import MenuBar from '../../components/MenuBar';
+import Wallet from './component/Wallet';
+import AddWalletDialog from './dialog/AddWalletDialog';
+import { AXIOS_METHOD, doApiCall, useApi } from '../../hooks/useApi';
+import { useAuth } from '../../hooks/useAuth';
 
-export default function WalletsScreen() {
+export default function WalletsScreen({ filterBy }) {
   const [open, setOpen] = useState(false);
   const [data, loading, error, forceWalletRefresh] = useApi(AXIOS_METHOD.GET, '/wallets', false);
+
   const { sessionUser } = useAuth();
+
+  function isOwner(name) {
+    return sessionUser.name === name;
+  }
 
   function handleOpen() {
     setOpen(true);
@@ -32,8 +36,6 @@ export default function WalletsScreen() {
     doApiCall(AXIOS_METHOD.DELETE, `/wallet/${id}`, onSuccess, () => handleClose());
   }
 
-  const isOwner = (name) => sessionUser.name === name;
-
   if (loading) {
     return <LinearProgress />;
   }
@@ -44,11 +46,11 @@ export default function WalletsScreen() {
 
   return (
     <Container maxWidth="lg">
-      <Grid container spacing={2}>
+      <Grid container spacing={3}>
         <Grid item lg={12} md={12} xs={12}>
           <MenuBar />
         </Grid>
-        {data && data.map((d) => (
+        {filterBy ? (data && data.filter(filterBy).map((d) => (
           <Wallet
             key={d.id}
             name={d.name}
@@ -58,9 +60,32 @@ export default function WalletsScreen() {
             forceWalletRefresh={forceWalletRefresh}
             isOwner={() => isOwner(d.created_by.name)}
           />
-        ))}
+        ))) : (data && data.map((d) => (
+          <Wallet
+            key={d.id}
+            name={d.name}
+            balance={d.balance}
+            id={d.id}
+            handleEvent={() => handleDelete(d.id)}
+            forceWalletRefresh={forceWalletRefresh}
+            isOwner={() => isOwner(d.created_by.name)}
+          />
+        )))}
         <Grid item lg={3} md={4} xs={12}>
-          <Button onClick={handleOpen} variant="contained" color="success" fullWidth sx={{ height: '175px' }} elevation={2}>
+          <Button
+            onClick={handleOpen}
+            variant="contained"
+            color="success"
+            fullWidth
+            sx={{
+              height: '200px',
+              transition: 'transform .5s, box-shadow 1s',
+              '&:hover': {
+                transform: 'scale(1.02) perspective(0px)',
+              },
+            }}
+            elevation={2}
+          >
             <Typography variant="h5">Add new wallet</Typography>
           </Button>
           <AddWalletDialog
